+++ src/components/MarkdownField.vue
@@ -51,7 +51,10 @@ function onKeydown(e: KeyboardEvent) {
   <div v-else>
     <div
       v-if="!editing"
-      class="min-h-[2em] px-2 py-1.5 text-xs bg-gray-800 rounded text-gray-200 cursor-text markdown-preview"
+      :class="[
+        'min-h-[2em] px-2 py-1.5 text-xs bg-gray-800 rounded text-gray-200 cursor-text markdown-preview',
+        !modelValue ? 'border border-gray-600' : ''
+      ]"
       @click="startEdit"
     >
       <div v-if="!modelValue" class="text-gray-500">Click to edit...</div>
@@ -121,7 +124,7 @@ function onKeydown(e: KeyboardEvent) {
 }
 .markdown-preview :deep(em) {
   font-style: italic;
-  color: rgb(156 163 175);
+  color: rgb(209 213 219);
 }
 
 .markdown-preview :deep(.dialogue) {